+++ changelog.md
@@ -1,2 +1,3 @@
 ##Additions/Improvements
-+ Fixes to various new V1.1 functions 
++ Fixes & compatibility improvements to all functions 
++ Added findAreaCircle function